% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitation-labelled-ms2.R
\name{quantify_labelled_ms2}
\alias{quantify_labelled_ms2}
\title{Labelled MS2 Quantitation}
\usage{
quantify_labelled_ms2(x, reporters, ...)
}
\arguments{
\item{x}{An instance of class \link{Spectra} with MS2 spectra.}

\item{reporters}{An instance of class \link{ReporterIons}.}

\item{...}{Additional parameters passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}
}
\value{
A instance of class \link{QFeatures} with as many assays as
there where acquisitions in \code{x}.
}
\description{
Function that converses a \code{Spectra} instance into an object of
class \code{QFeatures} by quantifying reporter ions of the MS2 spectra.
}
\author{
Laurent Gatto
}
