% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitation-labelled-ms3.R
\name{quantify_labelled_ms3}
\alias{quantify_labelled_ms3}
\title{Labelled MS3 Quantitation}
\usage{
quantify_labelled_ms3(x, reporters, ...)
}
\arguments{
\item{x}{An instance of class \link{Spectra} with MS2 and MS3 spectra.}

\item{reporters}{An instance of class \link{ReporterIons}.}

\item{...}{Additional parameters passed to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}.}
}
\value{
A instance of class \link{QFeatures} with as many assays as
there where acquisitions in \code{x}.
}
\description{
Function that converses a \code{Spectra} instance into an object of
class \code{QFeatures} by quantifying reporter ions of the MS3
spectra. The assays rowData are pulled from the respective MS2
precursor scans.
}
\examples{

basename(f <- msdata::proteomics(pattern = "MS3TMT11", full.names = TRUE))
x <- Spectra(f)

quantify_labelled_ms3(x, TMT11)
}
\author{
Laurent Gatto
}
