% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantParam.R
\docType{class}
\name{QuantParam}
\alias{QuantParam}
\alias{QuantParam-class}
\title{Quantiation Parameters}
\usage{
QuantParam(msLevel = integer(), params = list())
}
\arguments{
\item{msLevel}{\code{integer(1)} indicating the MS levels to be
quantified.}

\item{params}{\code{list()} containing additional method-specific
quantitation parameters.}
}
\description{
Quantitation of mass spectrometry data is implemented through the
\code{\link[=quantify]{quantify()}} method, that converts raw data objects such as
\code{\link[=Spectra]{Spectra()}} into \code{\link[=QFeatures]{QFeatures()}} objects that store and hande
quantitative data. Quantitation methods for MS data a numerous and
varied, and the exact method to be applied on the raw data is
paramterised by a \code{QuantParam} instance, i.e. a dedicated class
that stores all the parameters needed to perform a specific
quantitation method.
}
\section{Slots}{

\describe{
\item{\code{msLevel}}{\code{integer(1)} indicating the MS level the quantitation
will be performed on.}

\item{\code{params}}{named \code{list()} of additional parameters.}
}}

\examples{

## The default (empty) parameters
QuantParam()

## MS2 quantitation using TMT10 plex
QuantParam(msLevel = 2L, params = list(reporters = TMT10))
}
\author{
Laurent Gatto
}
