% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReporterIons-class.R, R/ReporterIons.R
\docType{class}
\name{ReporterIons}
\alias{ReporterIons}
\alias{ReporterIons-class}
\alias{show,ReporterIons-method}
\alias{[,ReporterIons,ANY,ANY,ANY-method}
\alias{length,ReporterIons-method}
\alias{mz,ReporterIons-method}
\alias{width,ReporterIons-method}
\alias{reporterNames}
\alias{reporterNames<-}
\alias{description,ReporterIons-method}
\alias{names,ReporterIons-method}
\alias{iTRAQ4}
\alias{iTRAQ5}
\alias{iTRAQ8}
\alias{iTRAQ9}
\alias{TMT10HCD}
\alias{TMT10}
\alias{TMT10ETD}
\alias{TMT11HCD}
\alias{TMT11}
\alias{TMT16HCD}
\alias{TMT16}
\alias{TMT6}
\alias{TMT6b}
\alias{TMT7}
\alias{TMT7b}
\title{Reporter Ions Class}
\format{
An object of class \code{ReporterIons} of length 4.

An object of class \code{ReporterIons} of length 5.

An object of class \code{ReporterIons} of length 8.

An object of class \code{ReporterIons} of length 9.

An object of class \code{ReporterIons} of length 10.

An object of class \code{ReporterIons} of length 10.

An object of class \code{ReporterIons} of length 10.

An object of class \code{ReporterIons} of length 11.

An object of class \code{ReporterIons} of length 11.

An object of class \code{ReporterIons} of length 16.

An object of class \code{ReporterIons} of length 16.

An object of class \code{ReporterIons} of length 6.

An object of class \code{ReporterIons} of length 6.

An object of class \code{ReporterIons} of length 7.

An object of class \code{ReporterIons} of length 7.
}
\usage{
ReporterIons(name, reporterNames, description = "", mz, width)

\S4method{show}{ReporterIons}(object)

\S4method{[}{ReporterIons,ANY,ANY,ANY}(x, i = "numeric", j = "missing", drop = "missing")

\S4method{length}{ReporterIons}(x)

\S4method{mz}{ReporterIons}(object)

\S4method{width}{ReporterIons}(x)

reporterNames(x)

reporterNames(x) <- value

\S4method{description}{ReporterIons}(object)

\S4method{names}{ReporterIons}(x)

iTRAQ4

iTRAQ5

iTRAQ8

iTRAQ9

TMT10HCD

TMT10

TMT10ETD

TMT11HCD

TMT11

TMT16HCD

TMT16

TMT6

TMT6b

TMT7

TMT7b
}
\arguments{
\item{name}{Parameter to set the \code{name} slot.}

\item{reporterNames}{Parameter to set the \code{reporterNames} slot.}

\item{description}{Parameter to set the \code{description}
slot. Default is \code{""}.}

\item{mz}{Parameter to set the \code{mz} slot.}

\item{width}{Parameter to set the \code{width} slot.}

\item{object}{An instance of class \code{ReporterIons}.}

\item{x}{An instance of class \code{ReporterIons}.}

\item{i}{\code{numeric()} to subset \code{x}.}

\item{j}{ignored.}

\item{drop}{ignored.}

\item{value}{A value for replacement.}
}
\description{
The \code{ReporterIons} class defines a set of isobaric reporter ions
that are used for quantification of labelled MS2 data, e.g. iTRAQ
(isobaric tag for relative and absolute quantitation) or TMT
(tandem mass tags).

Many commercial reporter ions are readily available. Custom
instances can be created with the \code{ReporterIons()} function.
}
\section{Slots}{

\describe{
\item{\code{name}}{\code{character(1)} naming the instance.}

\item{\code{reporterNames}}{\code{character()} of length equal to the number of
reporter ions in the instance. Used to uniquely name each ion.}

\item{\code{description}}{\code{character(1)} with a free-text description of
the instance. Default is an empty character.}

\item{\code{mz}}{\code{numeric()} with the reporter ions' m/z values.}

\item{\code{width}}{\code{numeric(1)} defining the region where the reporter
ion can be expected. This region is calculated as \code{mz} +/-
\code{width} for earch reporter ion.}
}}

\examples{

## Manual construction of a ReporterIons instance:
ReporterIons(description = "4-plex iTRAQ",
             name = "iTRAQ4",
             reporterNames = c("iTRAQ4.114", "iTRAQ4.115",
                               "iTRAQ4.116", "iTRAQ4.117"),
             mz = c(114.1, 115.1, 116.1, 117.1),
             width = 0.05)

## Some pre-defined reporter ions constructors:
TMT10

iTRAQ4

TMT10[1:5]

reporterNames(TMT10)

names(TMT10)

mz(TMT10)

width(TMT10)

as(TMT10, "DataFrame")
}
\references{
Ross PL, Huang YN, Marchese JN, Williamson B, Parker K, Hattan S,
Khainovski N, Pillai S, Dey S, Daniels S, Purkayastha S, Juhasz P,
Martin S, Bartlet-Jones M, He F, Jacobson A, Pappin
DJ. \emph{Multiplexed protein quantitation in Saccharomyces cerevisiae
using amine-reactive isobaric tagging reagents.} Mol Cell
Proteomics, 2004 Dec;3(12):1154-69. Epub 2004 Sep 22. PubMed
PMID: 15385600.

Thompson A, Schäfer J, Kuhn K, Kienle S, Schwarz J, Schmidt G,
Neumann T, Johnstone R, Mohammed AK, Hamon C. \emph{Tandem mass tags: a
novel quantification strategy for comparative analysis of complex
protein mixtures by MS/MS.} Anal Chem. 2003 Apr
15;75(8):1895-904. Erratum in: Anal Chem. 2006 Jun
15;78(12):4235. Mohammed, A Karim A and Anal Chem. 2003 Sep
15;75(18):4942. Johnstone, R. PubMed PMID: 12713048.
}
\author{
Laurent Gatto
}
\keyword{datasets}
