% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-methods.R
\name{quantify,Spectra-method}
\alias{quantify,Spectra-method}
\title{Quantitation of Raw MS Data}
\usage{
\S4method{quantify}{Spectra}(object, param, ...)
}
\arguments{
\item{object}{An instance of class \code{\link[=Spectra]{Spectra()}}.}

\item{param}{An instance of class \code{\link[=QuantParam]{QuantParam()}} that defines the
quantitation method.}
}
\value{
An instance of \code{\link[=QFeatures]{QFeatures()}}.
}
\description{

}
