% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification-labelled-ms2.R
\name{quantify_labelled_ms2}
\alias{quantify_labelled_ms2}
\title{Labelled MS2 Quantitation}
\usage{
quantify_labelled_ms2(x, reporters, ...)
}
\arguments{
\item{x}{An instance of class \link{Spectra} with MS2 spectra.}

\item{reporters}{An instance of class \code{\link[=ReportersIons]{ReportersIons()}}.}
}
\value{
A instance of class \link{QFeatures} with as many assays as
there where acquisitions in \code{x}.
}
\description{
Function that converses a \code{Spectra} instance into an object of
class \code{QFeatures} by quantifying reporter ions of the MS2 spectra.
}
\examples{
## Test data from the msdata package
f <- msdata::proteomics(pattern = "01.mzML.gz", full.names = TRUE) 
rw <- Spectra(f)

quantify_labelled_ms2(rw, TMT6)

## Simulate data from 2 files
rw <- filterMsLevel(rw, 2L) 
rw <- setBackend(rw, MsBackendDataFrame())

rw$dataOrigin <- sample(c("file1", "file2"),
                        length(rw),
                        replace = TRUE)

quantify_labelled_ms2(rw, TMT6)
}
\author{
Laurent Gatto
}
