% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantParam.R
\docType{class}
\name{QuantParam}
\alias{QuantParam}
\alias{QuantParam-class}
\alias{isEmpty,QuantParam-method}
\title{Quantiation Parameters}
\usage{
\S4method{isEmpty}{QuantParam}(x)

QuantParam(msLevel = NA_integer_, label = NA, params = list())
}
\arguments{
\item{x}{A instance of class \code{QuantParam}.}

\item{msLevel}{\code{integer(1)} indicating the MS levels to be
quantified. Default is \code{NA}.}

\item{label}{\code{logical(1)} defining is labelled or label-free
quantitation. Default is \code{NA}.}

\item{params}{\code{list()} containing additional method-specific
quantitation parameters. The list must be named.}
}
\description{
Quantitation of mass spectrometry data is implemented through the
\code{\link[=quantify]{quantify()}} method, that converts raw data objects such as
\code{\link[=Spectra]{Spectra()}} into \code{\link[=QFeatures]{QFeatures()}} objects that store and hande
quantitative data. Quantitation methods for MS data a numerous and
varied, and the exact method to be applied on the raw data is
paramterised by a \code{QuantParam} instance, i.e. a dedicated class
that stores all the parameters needed to perform a specific
quantitation method.

MS quantitation can be performed at the MS1 or MS2 (and 3)
levels. Additionally, quantitation can be label-free or
labelled. These two states are defined by the \code{msLevel} (an
integer) and \code{label} (a logical) parameters.

Additional details about the quantitation method are defined as a
named list, as described in the 'Details' section. The example
below illustrates how to defined labelled (\code{label = TRUE}) MS2
(\verb{msLevel = 2L) quantitation using TMT 10-plex isobaric tagging (defined by the }TMT10` \code{\link[=ReporterIons]{ReporterIons()}} instance).
}
\details{
Labelled MS2 Quantitation:
\itemize{
\item The \code{params} need only to contain a single mandatory element, a
\code{\link[=ReporterIons]{ReporterIons()}} instance, named \code{reporters}.
}

Labelled MS3 Quantitation:
\itemize{
\item The \code{params} need to contain a mandatory element, a
\code{\link[=ReporterIons]{ReporterIons()}} instance, named \code{reporters}.
\item Two optional character strings, named \code{acquisitionNum} and
\code{precScanNum}, containing the respective column names of the
acquition number and precursor scan acquisition numbers in the
raw data to be quantified. The default values are
\code{"acquisitionNum"} and \code{"precScanNum"} respectively, which match
the names in standard \code{Spectra} objects.
}
}
\section{Slots}{

\describe{
\item{\code{msLevel}}{\code{integer(1)} indicating the MS level the quantitation
will be performed on.}

\item{\code{label}}{}

\item{\code{params}}{named \code{list()} of additional parameters.}
}}

\examples{

## default (empty) parameters
QuantParam()

isEmpty(QuantParam())

## MS2 quantitation using TMT10 plex
QuantParam(msLevel = 2L,
           label = TRUE,
           params = list(reporters = TMT10))

## MS3 quantitation using TMT16 plex and non-standard acquisition
## number and precursor scan acqusition number
QuantParam(msLevel = 2L, label = TRUE,
           params = list(reporters = TMT10,
                         acqusitionNum = "aquisition_number",
                         preScanNum = "prec_scan_number"))
}
\author{
Laurent Gatto
}
